Do not run `gem update --system` on travis CI
Since it breaks the build for rubies < 2.3 because it tries to install
version 3 of rubygems, which is incompatible.

We could go around and keep updating system gems by trying to install
rubygems 3 and if it fails we try to install version 2, as stated in
rubygems issue #2534. But since most of the issues that made us always
update the system gems are solved (conflit between rubygems and
bundler) I saw no need for it

Related:
- Kinda like: plataformatec/devise@6e4b0b94bcec39c7dfdcc89e1f5768ab3e71defb
- Issue about the error: rubygems/rubygems#2534