Ensure `default` works as expected when used with `in`
`in` is a shortcut for declaring a scope with `as` and `using`, in other
words, these are essentially the same declarations:

    has_scope :title, as: :q, using: [:title]
    has_scope :title, in: :q

But the behavior wasn't always the same, since that is not clear. For
example when setting a default, you'd think this would work:

    has_scope :title, as: :q, using: [:title], default: { title: 'foo' }
    has_scope :title, in: :q, default: 'foo'

But since they're essentially the same, the latter default wouldn't be
applied, you'd have to set a default like the former: `{ title: 'foo' }`.

Now we make that conversion automatically for `in` to keep the reality
matching the expectations there.